refactor: update title with redis
